Clippy fixes.

+++ src/solver/mus_game.rs
@@ -134,7 +134,7 @@ impl MusGame {
     }
 
     pub fn default_utility_table() -> [[f64; 40]; 40] {
-        std::array::from_fn(|t1| std::array::from_fn(|t2| (t1 - t2) as f64))
+        std::array::from_fn(|t1| std::array::from_fn(|t2| t1 as f64 - t2 as f64))
     }
 
     fn iter_descartes<const N: usize>(
